# Conflicts:
#	src/main/java/DATN/dto/sanpham/SANPHAM_CREATE.java
#	src/main/java/DATN/dto/sanpham/SANPHAM_DETAILS.java

+++ src/main/java/DATN/dto/sanpham/SANPHAM_CREATE.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class SANPHAM_CREATE {
     // SAN_PHAM
     private String tensanpham;
-    private Integer dongia;
+    private long dongia;
     private Integer loai;
     private Integer thuonghieu;
     private String anhgoc;
